Comments resolved

+++ src/app/shared/components/configuration/advanced-settings/advanced-settings.component.ts
@@ -703,31 +703,36 @@ export class AdvancedSettingsComponent implements OnInit, OnDestroy {
     const that = this;
     that.isLoading = true;
     const valueField = this.skipExportForm.getRawValue();
-    if (this.showAddButton) {
+    if (this.showAddButton && this.expenseFilters.length > 1) {
       this.advancedSettingService
       .deleteExpenseFilter(this.expenseFilters[1].id)
       .subscribe((skipExport1: SkipExport) => {
       });
     }
-    if (!this.advancedSettingsForm.controls.skipExport.value) {
+    if (!this.advancedSettingsForm.controls.skipExport.value && this.expenseFilters.length > 0) {
       this.advancedSettingService
       .deleteExpenseFilter(this.expenseFilters[0].id)
       .subscribe((skipExport1: SkipExport) => {
       });
-      this.advancedSettingService
-      .deleteExpenseFilter(this.expenseFilters[1].id)
-      .subscribe((skipExport1: SkipExport) => {
-      });
+      if (this.expenseFilters.length > 1) {
+        this.advancedSettingService
+        .deleteExpenseFilter(this.expenseFilters[1].id)
+        .subscribe((skipExport1: SkipExport) => {
+        });
+      }
       that.isLoading = false;
     } else {
-    if (valueField.condition1.field_name !== 'report_title' && valueField.operator1 === 'iexact') {
-      valueField.operator1 = 'in';
-    }
-    if (valueField.join_by) {
-      if (valueField.condition2.field_name !== 'report_title' && valueField.operator2 === 'iexact') {
-        valueField.operator2 = 'in';
+      if (!valueField.condition1.field_name) {
+        return;
       }
-  }
+      if (valueField.condition1.field_name !== 'report_title' && valueField.operator1 === 'iexact') {
+        valueField.operator1 = 'in';
+      }
+      if (valueField.join_by) {
+        if (valueField.condition2.field_name !== 'report_title' && valueField.operator2 === 'iexact') {
+          valueField.operator2 = 'in';
+        }
+    }
     if (valueField.condition1.is_custom === true) {
       if (valueField.operator1 === 'is_empty') {
         valueField.value1 = ['True'];
